Rename commonTypes.h -> types.h
