+++ src/CMakeLists.txt
@@ -61,9 +61,10 @@ set(LIME_SUITE_SOURCES
     mcu_program/common_src/lms7002m_filters.c
 
     API/LMS_APIWrapper.cpp
-    include/limesuiteng/commonTypes.cpp
     include/limesuiteng/OpStatus.cpp
     include/limesuiteng/SDRDevice.cpp
+
+    utilities/toString.cpp
 )
 
 SET_SOURCE_FILES_PROPERTIES(
