+++ src/include/limesuiteng/SDRConfig.h
@@ -1,7 +1,7 @@
 #ifndef LIMESUITENG_SDRCONFIG_H
 #define LIMESUITENG_SDRCONFIG_H
 
-#include "limesuiteng/commonTypes.h"
+#include "limesuiteng/types.h"
 #include "limesuiteng/complex.h"
 #include "limesuiteng/GainTypes.h"
