+++ src/utilities/toString.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "limesuiteng/types.h"
+
+namespace lime {
+
+const char* ToCString(TRXDir dir);
+
+}
